Complete implementation of the scenario

+++ TravelPackage.java
@@ -2,9 +2,9 @@ class TravelPackage{
 
   String name;
   int price;
-  int rating;
+  float rating;
 
-  public TravelPackage(String name, int price, int rating){
+  public TravelPackage(String name, int price, float rating){
 
     this.name=name;
     this.price=price;
